show edit and remove buttons, and some contents  depending on the roles

+++ resources/views/livewire/category/category-details.blade.php
@@ -1,22 +1,24 @@
 <div>
 
     <b>{{ $category->title }}</b> <br>
-    <button wire:click="$emit('openModal', 'category.add-practice', {{ json_encode(['category_id' => $category->id ]) }})">Add Practice</button>
-    
+    @if($is_admin)
+        <button wire:click="$emit('openModal', 'category.add-practice', {{ json_encode(['category_id' => $category->id ]) }})">Add Practice</button>
+    @endif
     @foreach($practices as $practice)
         {{ $practice->id }} : {{ $practice->title }} : {{ $practice->description }} <br>
-
-        <button wire:click="edit_practice({{$practice->id}})" class="border-solid border-2 border-indigo-600 bg-green-400" :key="now() . $practice->id">EDIT</button>
-        <button wire:click.prevent="delete_practice({{$practice->id}})" class="border-solid border-2 border-indigo-600 bg-red-600">REMOVE</button> <br>
-
+        @if($is_admin)
+            <button wire:click="edit_practice({{$practice->id}})" class="border-solid border-2 border-indigo-600 bg-green-400" :key="now() . $practice->id">EDIT</button>
+            <button wire:click.prevent="delete_practice({{$practice->id}})" class="border-solid border-2 border-indigo-600 bg-red-600">REMOVE</button> <br>
+        @endif
     @endforeach
     <br><br><br>
+    @if($is_admin)
+        @foreach($users as $user)
+            {{ $user->id }} : {{ $user->name }}<br>
 
-    @foreach($users as $user)
-        {{ $user->id }} : {{ $user->name }}<br>
+            <button wire:click.prevent="delete_user({{$user->id}})" class="border-solid border-2 border-indigo-600 bg-red-600">REMOVE</button><br>
 
-        <button wire:click.prevent="delete_user({{$user->id}})" class="border-solid border-2 border-indigo-600 bg-red-600">REMOVE</button><br>
-
-    @endforeach
-    <button wire:click="$emit('openModal', 'category.add-user', {{ json_encode(['users' => $users, 'category_id' => $category->id ]) }})">Add User</button>
+        @endforeach
+        <button wire:click="$emit('openModal', 'category.add-user', {{ json_encode(['users' => $users, 'category_id' => $category->id ]) }})">Add User</button>
+    @endif
 </div>
